requirements.txt
flask>=2.0.0
openai>=1.0.0
faiss-cpu>=1.7.0
numpy>=1.20.0
pytz>=2021.1
tenacity>=8.0.0
tiktoken>=0.3.0
python-dotenv>=0.19.0
gunicorn>=20.1.0
google-cloud-storage
watchdog
filelock
requests
langgraph>=0.0.30
timedelta